+++ test/specs/data.js
@@ -18,7 +18,7 @@ describe('Initialization with data', function () {
     expect(fotorama.data.length).toEqual(3);
   });
 
-  it('can be initialized dinamically', function () {
+  it('can be initialized dinamically', function (done) {
     var api;
     $('<div class="gallery"></div>')
         .on('fotorama:ready', function (e, fotorama) {
@@ -35,11 +35,11 @@ describe('Initialization with data', function () {
 
     waitsFor(function () {
       return api;
-    }, 'Waits for fotorama initialization...', 100);
-
-    runs(function () {
+    }, function () {
       expect(api.data[0].img).toBe('test/i/okonechnikov/4-lo.jpg');
       expect($('.gallery .fotorama__wrap').width()).toBe(700);
+
+      done();
     });
   });