+++ test/specs/html.js
@@ -25,17 +25,17 @@ describe('HTML', function () {
     expect($('img', fotorama.activeFrame.$navThumbFrame).length).toBe(0);
   });
 
-  it('stage & nav image with the same `src` for the second frame `<div data-img="2-lo.jpg">2</div>`', function () {
+  it('stage & nav image with the same `src` for the second frame `<div data-img="2-lo.jpg">2</div>`', function (done) {
     fotorama.show({index: 1, time: 0});
 
     waitsFor(function () {
       return $('img', fotorama.activeFrame.$stageFrame).length;
-    }, 'Waits for the second image...', 100);
-
-    runs(function () {
+    }, function () {
       expect($('img', fotorama.activeFrame.$stageFrame).length).toBe(1);
       expect($('img', fotorama.activeFrame.$navThumbFrame).length).toBe(1);
       expect($('img', fotorama.activeFrame.$stageFrame).attr('src')).toBe($('img', fotorama.activeFrame.$navThumbFrame).attr('src'));
+
+      done();
     });
   });
 
@@ -43,16 +43,16 @@ describe('HTML', function () {
     expect(fotorama.activeFrame.html).toBe($('.fotorama__html > *', fotorama.activeFrame.$stageFrame)[0]);
   });
 
-  it('only thumb for the third frame `<div data-thumb="9-thumb.jpg">9</div>`', function () {
+  it('only thumb for the third frame `<div data-thumb="9-thumb.jpg">9</div>`', function (done) {
     fotorama.show({index: 2, time: 0});
 
     waitsFor(function () {
       return $('img', fotorama.activeFrame.$navThumbFrame).length;
-    }, 'Waits for the second image...', 100);
-
-    runs(function () {
+    }, function () {
       expect($('img', fotorama.activeFrame.$stageFrame).length).toBe(0);
       expect($('img', fotorama.activeFrame.$navThumbFrame).length).toBe(1);
+
+      done();
     });
   });
 
@@ -60,31 +60,31 @@ describe('HTML', function () {
     expect(fotorama.activeFrame.html).toBe($('.fotorama__html > *', fotorama.activeFrame.$stageFrame)[0]);
   });
 
-  it('different stage & nav images for the fourth frame `<div data-img="6-lo.jpg" data-thumb="6-thumb.jpg">6</div>`', function () {
+  it('different stage & nav images for the fourth frame `<div data-img="6-lo.jpg" data-thumb="6-thumb.jpg">6</div>`', function (done) {
     fotorama.show({index: 3, time: 0});
 
     waitsFor(function () {
       return $('img', fotorama.activeFrame.$stageFrame).length;
-    }, 'Waits for the second image...', 100);
-
-    runs(function () {
+    }, function () {
       expect($('img', fotorama.activeFrame.$stageFrame).length).toBe(1);
       expect($('img', fotorama.activeFrame.$navThumbFrame).length).toBe(1);
       expect($('img', fotorama.activeFrame.$stageFrame).attr('src')).not.toBe($('img', fotorama.activeFrame.$navThumbFrame).attr('src'));
+
+      done();
     });
   });
 
-  it('last test', function () {
+  it('last test', function (done) {
     fotorama.show('>>');
 
     waitsFor(function () {
       return fotorama.activeFrame.$stageFrame.data('state') === 'error';
-    }, 'wait for a state of the frame', 100);
-
-    runs(function () {
+    }, function () {
       expect($('img', fotorama.activeFrame.$stageFrame).length).toBe(0);
       expect($('img', fotorama.activeFrame.$navThumbFrame).length).toBe(0);
       expect($('.fotorama__html', fotorama.activeFrame.$stageFrame).text()).toBe('LAST');
+
+      done();
     });
   });
 });