+++ test/specs/video.js
@@ -19,12 +19,12 @@ describe('Video', function () {
   });
 
 
-  it('previews are fetched (works only if script able to connect to the Internet)', function () {
+  it('previews are fetched (works only if script able to connect to the Internet)', function (done) {
     $thumb.each(function () {
       var $this = $(this);
       waitsFor(function () {
         return $('.fotorama__img', $this)[0];
-      }, 'Here must be an image...', 1000);
+      }, done);
     });
   });