+++ HISTORY.md
@@ -2,7 +2,70 @@
 
 > https://github.com/artpolikarpov/fotorama/releases
 
-## 4.4.9, Dec 30 2013
+## 4.5.1, 29 March 2014
+
+<!--4.5.1:name-->
+
+Keep it ace
+
+<!--/4.5.1:name-->
+
+<!--4.5.1:notes-->
+
+* Fix animations and visibility of arrows and captions.
+* #181 Always visible arrows via `arrows:'always'`.
+* #214 Fullscreen mode now using complete width and surpass the `maxwidth` limit.
+* #218 Allow to hide captions via setOptions.
+* Make `loop:true` and `clicktransition:'crossfade'` play well together.
+* #219 Extended keyboard support.
+* Fix some measures collisions.
+* Optimize, test.
+
+### Extended keyboard support
+The old `data-keyboard="true"` enables keyboard navigation with the arrows. Extend it passing the object with keys you want to turn on:
+
+```javascript
+<div class="fotorama"
+     data-keyboard='{"space":true, "home":true, "end":true}'>
+  <img src="1.jpg">
+  <img src="2.jpg">
+</div>
+```
+
+<!--/4.5.1:notes-->
+
+## 4.5.0, 20 March 2014
+
+<!--4.5.0:name-->
+
+Yana
+
+<!--/4.5.0:name-->
+
+<!--4.5.0:notes-->
+
+New options:
+* `navwidth` — sets the special width of the nav;
+* `clicktransition` — alternative transition after click, now it is possible to slide after swipe and fade after click.
+
+Now fotorama shows only first raw slide while initialization.
+Class `.fotorama--unobtrusive` makes fotorama unobtrusive and shows all raw slides,
+use it when you care about users without JavaScript.
+
+Closed issues:
+
+* #159 Found the balance between swipe, click, and arrows.
+* #160 Fixed tiny drag on the arrows.
+* #174 Crush sprite files (via @zslabs).
+* #187 Fix compatibility with `$.noConflict(true)` (via @Lendar).
+* #192 Fixed `stopautoplayontouch`.
+* #205 Fixed a typo that forced using the http protocol (via @dhayab).
+
+Fixed `glimpse`, nav overflow bug, and other minors.
+
+<!--/4.5.0:notes-->
+
+## 4.4.9, 30 December 2013
 
 <!--4.4.9:name-->
 
@@ -21,7 +84,7 @@ Fixed issues:
 
 <!--/4.4.9:notes-->
 
-## 4.4.8, Nov 13 2013
+## 4.4.8, 13 November 2013
 
 <!--4.4.8:name-->
 
@@ -39,7 +102,7 @@ Five hot fixes
 
 <!--/4.4.8:notes-->
 
-## 4.4.7, Nov 5 2013
+## 4.4.7, 5 November 2013
 
 <!--4.4.7:name-->
 
@@ -56,7 +119,7 @@ Spin the spinner
 
 <!--/4.4.7:notes-->
 
-## 4.4.6, Oct 20 2013
+## 4.4.6, 20 October 2013
 
 <!--4.4.6:name-->
 
@@ -70,7 +133,7 @@ Optimizations
 
 <!--/4.4.6:notes-->
 
-## 4.4.5, Oct 17 2013
+## 4.4.5, 17 October 2013
 
 <!--4.4.5:name-->